+++ app/react/V2/Components/Metadata/MetadataRecord.tsx
@@ -1,13 +1,8 @@
-import React, { useEffect, useLayoutEffect, useMemo, useRef } from 'react';
-import { useAtomValue, useSetAtom } from 'jotai';
+import React, { useMemo, useRef } from 'react';
+import { useAtomValue } from 'jotai';
 import { Translate } from '#app/I18N/index.js';
 import { templatesAtom } from '#V2/atoms/templatesAtom.js';
 import { Entity } from '#V2/api/entities/types.js';
-import {
-  applyMetadataFieldFocus,
-  FLASH_MS,
-  focusMetadataFieldAtom,
-} from './focusMetadataFieldAtom.js';
 import {
   Date,
   RelationshipCards,
@@ -23,58 +18,17 @@ import { isLongField, partitionMetadataRecord } from './metadataPropertyLayout.j
 import { renderFieldContent, renderScalarContent } from './Components/metadataFieldContent.js';
 import { fieldTitle, specializedCardTitle } from './Components/metadataFieldTitle.js';
 import { connectionPillsForField, type OpenEntityTarget } from './Components/ConnectionPills.js';
+import { useMetadataRecordFocus } from './useMetadataRecordFocus.js';
 
 type MetadataRecordProps = {
   entity: Entity;
   onOpenEntity?: (target: OpenEntityTarget) => void;
 };
 
-// eslint-disable-next-line max-statements
 const MetadataRecord = ({ entity, onOpenEntity }: MetadataRecordProps) => {
   const templates = useAtomValue(templatesAtom);
-  const focusField = useAtomValue(focusMetadataFieldAtom);
-  const clearFocus = useSetAtom(focusMetadataFieldAtom);
   const rootRef = useRef<HTMLDivElement>(null);
-  const prevSharedIdRef = useRef(entity.sharedId);
-  const ownsFocusRef = useRef(false);
-
-  useEffect(() => {
-    if (prevSharedIdRef.current !== entity.sharedId) {
-      ownsFocusRef.current = false;
-      clearFocus(null);
-      prevSharedIdRef.current = entity.sharedId;
-    }
-  }, [entity.sharedId, clearFocus]);
-
-  useEffect(
-    () => () => {
-      if (ownsFocusRef.current) {
-        ownsFocusRef.current = false;
-        clearFocus(null);
-      }
-    },
-    [clearFocus]
-  );
-
-  useLayoutEffect(() => {
-    if (!focusField) return undefined;
-    let clearTimer: number | undefined;
-    const cleanup = applyMetadataFieldFocus(
-      () => rootRef.current,
-      focusField.fieldKey,
-      () => {
-        ownsFocusRef.current = true;
-        clearTimer = window.setTimeout(() => {
-          ownsFocusRef.current = false;
-          clearFocus(null);
-        }, FLASH_MS);
-      }
-    );
-    return () => {
-      cleanup();
-      if (clearTimer !== undefined) window.clearTimeout(clearTimer);
-    };
-  }, [focusField, clearFocus]);
+  useMetadataRecordFocus(entity.sharedId, rootRef);
 
   const { entityTemplate, metadata } = useFormatMetadata(entity, templates, {
     groupGeolocationProperties: true,
