+++ app/react/V2/Routes/Entity/Components/Files/FileDetailsReadonlyMeta.tsx
@@ -4,24 +4,6 @@ import { getRowIcon } from './fileRowIcon.js';
 import { FileDetailsField } from './FileDetailsField.js';
 import { EntityFileRow } from './types.js';
 
-const FileDetailsCard = ({
-  headerAction,
-  children,
-}: {
-  headerAction: React.ReactNode;
-  children: React.ReactNode;
-}) => (
-  <div className="space-y-3 rounded-md bg-warm p-4">
-    <div className="flex items-center justify-between">
-      <div className="text-xs font-semibold uppercase tracking-wider text-ink-tertiary">
-        <Translate>File details</Translate>
-      </div>
-      {headerAction}
-    </div>
-    {children}
-  </div>
-);
-
 const FileDetailsReadonlyMeta = ({ row }: { row: EntityFileRow }) => (
   <>
     <FileDetailsField label={<Translate>Type</Translate>}>
@@ -39,4 +21,4 @@ const FileDetailsReadonlyMeta = ({ row }: { row: EntityFileRow }) => (
   </>
 );
 
-export { FileDetailsCard, FileDetailsReadonlyMeta };
+export { FileDetailsReadonlyMeta };
